refactor: relocate dynamic orchestration to provider context and decouple synchronized wrapper
