+++ src/client/mod.ts
@@ -5,8 +5,5 @@ export * from "./sparql-engine/mod.ts";
 export * from "./search-index/mod.ts";
 export * from "./providers/rdfjs/mod.ts";
 export * from "./providers/comunica/mod.ts";
-export * from "./factory.ts";
-export {
-  createClient as createLibsqlClient,
-  type CreateClientOptions,
-} from "./providers/libsql/libsql-client.ts";
+
+export * from "./providers/libsql/libsql-provider.ts";
