+++ src/client/providers/libsql/mod.ts
@@ -1,4 +1,4 @@
 export * from "./libsql-search-index.ts";
-export * from "./libsql-client.ts";
+export * from "./libsql-provider.ts";
 export * from "./libsql-quad-hydrator.ts";
 export * from "./libsql-quad-synchronizer.ts";
